requirements.txt
streamlit>=1.28.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=1.0.0
requests>=2.31.0
pandas>=2.0.0
pysqlite3-binary
dropbox
python-dotenv
Pillow>=9.0.0